test by Abdulbosit

+++ src/main/java/uz/pdp/backend/olxapp/service/CategoryServiceImpl.java
@@ -1,8 +1,10 @@
 package uz.pdp.backend.olxapp.service;
 
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import uz.pdp.backend.olxapp.entity.Category;
 import uz.pdp.backend.olxapp.exception.ConflictException;
 import uz.pdp.backend.olxapp.exception.EntityNotFoundException;
@@ -23,7 +25,7 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public List<CategoryDTO> getAllCategories() {
 
-        List<Category> categories = categoryRepository.findAll();
+        List<Category> categories = categoryRepository.getCategory();
         return categories.stream().map(categoryMapper::toDto).toList();
 
     }
@@ -37,6 +39,7 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
+    @Transactional
     public CategoryDTO save(CategoryReqDTO categoryReqDTO) {
 
         Category parentCategory = null;
@@ -62,7 +65,8 @@ public class CategoryServiceImpl implements CategoryService {
 
 
     @Override
-    public CategoryDTO update(Long id, CategoryDTO categoryDTO) {
+    @Transactional
+    public CategoryDTO update(Long id, @Valid CategoryReqDTO categoryDTO) {
 
         Category category = categoryRepository.findByIdOrThrow(id);
 
@@ -74,9 +78,14 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
+    @Transactional
     public CategoryDTO delete(Long id) {
 
         Category category = categoryRepository.findByIdOrThrow(id);
+
+        if (!category.getChildren().isEmpty()) {
+            throw new EntityNotFoundException("Can't delete category with child categories!", HttpStatus.CONFLICT);
+        }
 
         categoryRepository.delete(category);
         return categoryMapper.toDto(category);
